% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corona_cont.R
\docType{data}
\name{corona_cont}
\alias{corona_cont}
\title{corona_cont}
\format{
a tibble with 2,880 observations and 4 variables
\itemize{
\item \strong{date}: Date of recording
\item \strong{continent}: full continent name
\item \strong{type}: type of coronavirus case being confirmed, death or recovered
\item \strong{cases}: number of cases
}
}
\source{
\href{https://github.com/RamiKrispin/coronavirus}{Coronavirus Github}
}
\usage{
corona_cont
}
\description{
This data set shows the daily coronavirus cases by each continent by case type
}
\keyword{datasets}
