% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoronaShiny-package.R
\docType{package}
\name{CoronaShiny-package}
\alias{CoronaShiny}
\alias{CoronaShiny-package}
\title{CoronaShiny: Communicating with Data Coronavirus Shiny Application Package}
\description{
This package contains the shiny application exploring the current state of coronavirus.
}
\author{
\strong{Maintainer}: Vinny Vu \email{vvuu0009@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
