% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_july20.R
\docType{data}
\name{coronavirus_july20}
\alias{coronavirus_july20}
\title{56mins}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 188 rows and 5 columns.
}
\usage{
coronavirus_july20
}
\description{
This shows cumulative cases as at July 2020
}
\keyword{datasets}
