% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Running Shiny Application}
\usage{
launch_app()
}
\value{
The Shiny Application exploring the state of Coronavirus as at July 31 2020
}
\description{
This function runs the Coronavirus Shiny Application exploring the state of Coronavirus as at July 31 2020
}
