% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInput01.R
\name{selectInput01}
\alias{selectInput01}
\title{Running selectInput for corona app}
\usage{
selectInput01(id, column)
}
\description{
This function changes the selectInput function of the Shiny app based on the desired input

@export
}
