% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_july20.R
\docType{data}
\name{coronavirus_july20}
\alias{coronavirus_july20}
\title{coronavirus_july20}
\format{
a tibble with 188 observations and 5 variables
\itemize{
\item \strong{Country}: fully country name
\item \strong{Continent}: full continent name
\item \strong{Total Confirmed}: the cumulative confirmed coronavirus cases as at 31 July 2020
\item \strong{Total Deaths}: the cumulative coronavirus death cases as at 31 July 2020
\item \strong{Total Recovered}: the cumulative recovery coronavirus case as at 31 July 2020
}
}
\source{
\href{https://github.com/RamiKrispin/coronavirus}{Coronavirus Github}
}
\usage{
coronavirus_july20
}
\description{
This data set shows the cumulative coronavirus cases as at 31 July 2020 for each country
}
\keyword{datasets}
