% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInput01.R
\name{selectInput01}
\alias{selectInput01}
\title{Running selectInput for corona app}
\usage{
selectInput01(id, column)
}
\arguments{
\item{id}{The variable we wish to include in the selectInput}

\item{column}{The specific column from the data set we wish to extract the values to appear as the options for the selectInput}
}
\value{
selectInput("continent", "select the continent", choices = corona_cont$continent)
}
\description{
This function changes the selectInput function of the Shiny app based on the desired input
}
\examples{
selectInput01("continent", corona_cont$continent)

}
